requirements.txt
openai>=1.12.0
langgraph>=0.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pandas>=2.2.0
PyYAML>=6.0.2